Semantic Line Breaks, +more
Reworded the opening paragraph of the CoM. Use of 'they'
was ambiguous, and the plurality of 'solutions' did not match
the verb 'is'.
Also removed the 'forever' from statement 4. Adding 'forever'
makes it seem as though one may expect others to do work
for him in the short term.
Removed 'world' from statement 7. Politics in general should not
be part of a technical discussion, not just world politics.